Add a script for renaming the bundle and test it in CI
This used to live in the https://gitlab.tugraz.at/dbp/cli/ for creating a new
bundle.

To make this more testable and keep the template code and the renaming logic close
we include the renaming part here and also test in CI that it creates a working bundle.

The next step is to change the CLI, so it onyl clones the project, then renames using this
script and then removes unneeded files.